feat(filters): add write filters
