+++ generator/templates/query.gotpl
@@ -297,31 +297,30 @@
 			{{ end }}
 
 			{{ if not $field.IsList }}
-				{{ range $type := $.DMMF.WriteTypes }}
-					{{ if eq $field.Type $type.Name }}
-						{{ range $method := $type.Methods }}
-							// {{ $method.Name }} the {{ if $field.IsRequired }}required{{ else }}optional{{ end }} value of {{ $field.Name.GoCase }}
-							func (r {{ $struct }}) {{ $method.Name }}(value {{ $field.Type.Value }}) {{ $setReturnStruct }} {
-								return {{ $setReturnStruct }}{
-									data: builder.Field{
-										Name:   "{{ $field.Name }}",
-										Fields: []builder.Field{
-											builder.Field{
-												Name:   "{{ $method.Action }}",
-												Value:  value,
-											},
+				{{ $writeType := $.AST.WriteFilter $field.Type.String }}
+				{{ if $writeType }}
+					{{ range $method := $writeType.Methods }}
+						// {{ $method.Name }} the {{ if $field.IsRequired }}required{{ else }}optional{{ end }} value of {{ $field.Name.GoCase }}
+						func (r {{ $struct }}) {{ $method.Name }}(value {{ $field.Type.Value }}) {{ $setReturnStruct }} {
+							return {{ $setReturnStruct }}{
+								data: builder.Field{
+									Name:   "{{ $field.Name }}",
+									Fields: []builder.Field{
+										builder.Field{
+											Name:   "{{ $method.Action }}",
+											Value:  value,
 										},
 									},
-								}
+								},
 							}
+						}
 
-							func (r {{ $struct }}) {{ $method.Name }}IfPresent(value *{{ $field.Type.Value }}) {{ $setReturnStruct }} {
-								if value == nil {
-									return {{ $setReturnStruct }}{}
-								}
-								return r.{{ $method.Name }}(*value)
+						func (r {{ $struct }}) {{ $method.Name }}IfPresent(value *{{ $field.Type.Value }}) {{ $setReturnStruct }} {
+							if value == nil {
+								return {{ $setReturnStruct }}{}
 							}
-						{{ end }}
+							return r.{{ $method.Name }}(*value)
+						}
 					{{ end }}
 				{{ end }}
 			{{ end }}
@@ -423,9 +422,9 @@
 		{{ end }}
 
 		{{/* Provide field and type-specific methods. */}}
-		{{ $type := $.AST.ReadFilter $field.Type.String }}
-		{{ if $type }}
-			{{ range $method := $type.Methods }}
+		{{ $readType := $.AST.ReadFilter $field.Type.String }}
+		{{ if $readType }}
+			{{ range $method := $readType.Methods }}
 				func (r {{ $struct }}) {{ $method.Name }}(value {{ if $field.IsList }}[]{{ end }}{{ $field.Type.Value }}) {{ $returnStruct }} {
 					{{ if $field.IsList }}
 						var fields []builder.Field
